+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # CHANGELOG
 
-## 0.1
+## 0.2.0
+
+- Add `Attention` base class, `MultiHeadAttention`, and `ScaledDotProductAttention` classes.
+- Add `branch_network` and `trunk_network` arguments to `DeepONet` to allow for custom network architectures.
+- Add `MaskedOperator` base class.
+- Add `DeepCatOperator`.
+
+## 0.1.0
 
 - Move all content of `__init__.py` files to sub-modules.
 - Add `Trainer` class to replace `operator.fit` method.